añdo las funcionalidades basicas en todos

+++ src/users/users.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserDto, UpdateUserDto } from './user.dto';
@@ -9,7 +10,19 @@ export class UserService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    private readonly jwtService: JwtService,
   ) {}
+
+  async login(email: string, password: string): Promise<{ accessToken: string }> {
+    const user = await this.userRepository.findOne({ where: { email } });
+    if (!user || user.password !== password) {
+      throw new UnauthorizedException('Invalid credentials');
+    }
+
+    const payload = { email: user.email, name: user.name, sub: user.id };
+    const accessToken = this.jwtService.sign(payload, { secret: process.env.SECRET_KEY });
+    return { accessToken };
+  }
 
   async findAll(): Promise<User[]> {
     return this.userRepository.find();
@@ -20,16 +33,32 @@ export class UserService {
   }
 
   async create(createUserDto: CreateUserDto): Promise<User> {
+    const userFindEmail = await this.userRepository.findOne({ where: { email:createUserDto.email } });
+    if(userFindEmail){
+      throw new UnauthorizedException('Email already exists');
+    }
+    const userFindName = await this.userRepository.findOne({ where: { email:createUserDto.name } });
+    if(userFindName){
+      throw new UnauthorizedException('Username already exists');
+    }
     const user = this.userRepository.create(createUserDto);
     return this.userRepository.save(user);
   }
 
   async update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
+    const user = await this.userRepository.findOne({ where: { id } });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
     await this.userRepository.update(id, updateUserDto);
     return this.userRepository.findOne({ where: { id } });
   }
 
   async remove(id: number): Promise<void> {
+    const user = await this.userRepository.findOne({ where: { id } });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
     await this.userRepository.delete(id);
   }
 }
